Set special notices for deprecated or renamed plugins on old versions.
The idea is to help augment the automated "You are viewing an old page!"
notification. The naive approach I took assumed that there was always a
page with the same name in future versions (/docs/x.y.z/something) and
that wasn't true for things like the amqp-to-rabbitmq rename, the
elasticsearch_http deprecation, the grep deprecation, collectd input
moving to be a codec, etc.

The goal here is to provide a specific notification per page (where
relevant) and redirecting users to the best places.

Also bump to Logstash 1.5.1

Testing: I haven't deployed this yet, but having Firefox load this
javascript into logstash.net (via the web console) gets me the right
content I expect.

Helps to resolve https://github.com/elastic/logstash/issues/3359